beacon chain db to node.go instead of service_setup
